feat : make U-Net model.
I implemented model architecture in U-Net papers. In the case of crop,
the x and y coordinates were randomly selected from the shape of the
tensor (x1, x2, x3, x4), which was the previous process(=copy), minus
the shape of the sensor (x), and then the shape of the tensor (x) was
added to the selected x and y.